--- a/crates/polars-core/src/chunked_array/temporal/duration.rs
+++ b/crates/polars-core/src/chunked_array/temporal/duration.rs
@@ -1,4 +1,5 @@
 use crate::export::chrono::Duration as ChronoDuration;
+use crate::fmt::fmt_duration_string;
 use crate::prelude::DataType::Duration;
 use crate::prelude::*;
 
@@ -58,6 +59,24 @@ impl DurationChunked {
     /// Change the underlying [`TimeUnit`]. This does not modify the data.
     pub fn set_time_unit(&mut self, tu: TimeUnit) {
         self.2 = Some(Duration(tu))
+    }
+
+    /// Convert from [`Duration`] to String; note that `strftime` format
+    /// strings are not supported, only the specifiers 'iso' and 'polars'.
+    pub fn to_string(&self, format: &str) -> PolarsResult<StringChunked> {
+        match format {
+            "iso" | "polars" => {
+                let out: StringChunked = self
+                    .0
+                    .apply_nonnull_values_generic(DataType::String, |v: i64| {
+                        fmt_duration_string(v, self.time_unit(), format == "iso")
+                    });
+                Ok(out)
+            },
+            _ => Err(PolarsError::InvalidOperation(
+                format!("format {:?} not supported for Duration type (expected one of 'iso' or 'polars')", format).into(),
+            )),
+        }
     }
 
     /// Construct a new [`DurationChunked`] from an iterator over [`ChronoDuration`].